requirements.txt
aiohttp==3.7.4.post0
beautifulsoup4==4.11.2
humanize==4.5.0
numpy==1.23.4
pandas==1.5.1
pycountry==22.3.5
python_dateutil==2.8.2
PyYAML==6.0
scikit_learn==1.2.1
scipy==1.10.0
selenium==4.8.0
SQLAlchemy==1.4.44
tabulate==0.9.0
termcolor==2.2.0
tqdm==4.64.0
lxml
psycopg2
ipywidgets
ipykernel
jupyterlab
iteround
